Added resolve pass and extra imgui rendering pass:
Needed to add extra rendering pass with compatible framebuffer to solve
this, which boils down to creating a rendering pass with LOAD_DONT_CARE
that is compatible with the color pass. I also needed to transition the
frambuffer we're going to use to the right layout for the resolve pass,
that boils down to: VK_IMAGE_LAYOUT_GENERAL for VK_STORAGE_IMAGE stuff.